Index resorted and experiments with annotater
Sorry, but I am really getting annoyed if something is not working and waste hours and hours. But I think the annotater thing is going above my capacities. I guess that it has something to do with annotater.js (https://github.com/beast911/annotater/tree/master/src) and that the custom javascript code has to be included in jekyll as an asset (https://github.com/envygeeks/jekyll-assets/blob/master/README.md). But this involves too many changes in the configuration and I would probably break something if I would try it.