ontology page
Presentations and guidance documents have been added from eNanoMapper, NanoCommons and ACEnano. However, some general problems are still remaining: 1) Putting pdfs into presentation mode does not work. This might be a problem with the google pdf viewer but it is still annoying since the embedded pdfs are nice to have a quick view but they are too small to read the text in the slides. Making the embeddings bigger to be able to read all the text is not really an opion since it would be quite complicated to find anything with all the big iframes. 2) It would be great to have an easiy way to automatically make the height of the embedded pdf to show the full first page and only this. 3) Because of many reasions, we put many documents on zenodo. However, I did not find an easy way to have a preview of the zenodo page or the document in the user handbook. Zenodo doesn't allow to show its webpages as iframes and the documents are only available as downloads. As simple screenshot would be enough but I could not find a way to automate this as this is available in wordpress.